+++ tests/fixtures/css/at_rules/media_query_simple/input_invalid_media_no_query_or_block.svelte
@@ -1,0 +1,3 @@
+<style>
+	@media
+</style>