+++ tests/fixtures/svelte/blocks/snippet/basic/input_invalid_snippet_no_name.svelte
@@ -1,0 +1,1 @@
+{#snippet}a{/snippet}